refactor: Choose different character for new separator
This avoids problems when updating from an older version. With "|", which was included in the old separator "<|>",
the update of the serialization format won't be detected. Instead, the fields would have stray "<" and ">" added.
Choosing "$", a completely unrelated character, deserialization of the older format will likely fail and an appropriate
error message can be displayed.